Change Setting into Toggle component
