--- src/components/Setting/index.js
@@ -1,3 +1,0 @@
-import Setting from './Setting';
-
-export default Setting;
