+++ src/components/Toggle/Toggle.js
@@ -1,91 +1,86 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled, { css } from 'react-emotion';
-import { size } from 'polished';
 
-const TRACK_WIDTH = 40;
-const TRACK_HEIGHT = 24;
-const KNOB_SIZE = 16;
-const ANIMATION_TIMING = '200ms ease-in-out';
+import { Switch } from './components';
+import Text from '../Text';
 
-const knobShadow = color => `
-  0 2px 0 0 ${color}, inset 0 2px #FFF,
-    inset 0 1px 0 1px rgba(255, 255, 255, 6%)
+const textWrapperStyles = ({ theme }) => css`
+  label: toggle__text-wrapper;
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+  margin-left: ${theme.spacings.kilo};
 `;
 
-const trackBaseStyles = ({ theme }) => css`
-  label: toggle;
-  background-color: ${theme.colors.n300};
-  border-radius: ${TRACK_HEIGHT}px;
-  position: relative;
-  transition: background-color ${ANIMATION_TIMING};
-  ${size(TRACK_HEIGHT, TRACK_WIDTH)};
-`;
-
-const trackOnStyles = ({ theme, on }) =>
-  on &&
-  css`
-    label: toggle--on;
-    background-color: ${theme.colors.b500};
-  `;
-
-const ToggleTrack = styled('div', { label: 'ToggleTrack' })(
-  trackBaseStyles,
-  trackOnStyles
+const ToggleTextWrapper = styled('div', { label: 'ToggleTextWrapper' })(
+  textWrapperStyles
 );
 
-const knobBaseStyles = ({ theme }) => css`
-  label: toggle__knob;
-  background-color: ${theme.colors.n100};
-  box-shadow: ${knobShadow(theme.colors.n500)};
-  position: absolute;
-  top: 50%;
-  transform: translate3d(${theme.spacings.bit}, -50%, 0);
-  transition: box-shadow ${ANIMATION_TIMING}, transform ${ANIMATION_TIMING};
-  ${size(KNOB_SIZE)};
-  border-radius: ${KNOB_SIZE}px;
+const labelStyles = css`
+  label: toggle__label;
+  vertical-align: 1px;
 `;
 
-const knobOnStyles = ({ theme, on }) =>
-  on &&
+const ToggleLabel = styled(Text, { label: 'ToggleLabel' })(labelStyles);
+ToggleLabel.propTypes = Text.propTypes;
+ToggleLabel.defaultProps = Text.defaultProps;
+
+const explanationStyles = ({ theme }) => css`
+  label: toggle__explanation;
+  color: ${theme.colors.n500};
+`;
+
+const ToggleExplanation = styled(Text, { label: 'ToggleExplanation' })(
+  explanationStyles
+);
+ToggleExplanation.propTypes = Text.propTypes;
+ToggleExplanation.defaultProps = Text.defaultProps;
+
+const toggleWrapperStyles = css`
+  label: toggle;
+  display: flex;
+  flex-align: flex-start;
+`;
+
+const toggleWrapperMarginStyles = ({ theme, withMargin }) =>
+  withMargin &&
   css`
-    label: toggle__knob--on;
-    box-shadow: ${knobShadow(theme.colors.b700)};
-    transform: translate3d(
-      calc(${TRACK_WIDTH - KNOB_SIZE}px - ${theme.spacings.bit}),
-      -50%,
-      0
-    );
+    label: toggle--with-margin;
+    margin-bottom: ${theme.spacings.mega};
   `;
 
-const ToggleKnob = styled('div', { label: 'ToggleKnob' })(
-  knobBaseStyles,
-  knobOnStyles
+const ToggleWrapper = styled('div', { label: 'ToggleWrapper' })(
+  toggleWrapperStyles,
+  toggleWrapperMarginStyles
 );
 
 /**
- * A simple Toggle component.
+ * Describe your component here.
  */
-const Toggle = ({ on, onToggle }) => (
-  <ToggleTrack onClick={onToggle} on={on} role="switch" aria-checked={on}>
-    <ToggleKnob {...{ on }} />
-  </ToggleTrack>
+const Toggle = ({ label, explanation, withMargin, ...props }) => (
+  <ToggleWrapper {...{ withMargin }}>
+    <Switch {...props} />
+    <ToggleTextWrapper>
+      <ToggleLabel element="label" size="kilo">
+        {label}
+      </ToggleLabel>
+      {explanation && (
+        <ToggleExplanation size="kilo">{explanation}</ToggleExplanation>
+      )}
+    </ToggleTextWrapper>
+  </ToggleWrapper>
 );
 
 Toggle.propTypes = {
-  /**
-   * Is the toggle on?
-   */
-  on: PropTypes.bool,
-  /**
-   * Toggle callback used as onClick.
-   */
-  onToggle: PropTypes.func
+  label: PropTypes.string.isRequired,
+  withMargin: PropTypes.bool,
+  explanation: PropTypes.string
 };
 
 Toggle.defaultProps = {
-  on: false,
-  onToggle: () => {}
+  withMargin: true,
+  explanation: null
 };
 
 /**
