+++ src/components/Toggle/components/index.js
@@ -1,0 +1,3 @@
+import Switch from './Switch';
+
+export { Switch };
